requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
requests
python-decouple
duckduckgo-search
